First version of lab1 - 2020/09/28
Created file, MUX_HEXDEC, fairly simple, 2:1 mux that will either output the hex values, or the BCD values to the 7 segment displays.

Still need to create an instance of the mux in top_level, reassign/assign pins, and create a test bench file/run simulations.

all modifications will be labeled as:

--MT START MODS
--MT END MODS

FYI
"--" is the comment syntax for VHDL