Updated Project Files + Compilation
Added all used files into lab2 folder so lab2 folder could be a standalone project, easier when cloning from online.

Project compiled, using the same pin assignments as lab 1.

left out any files in this push that were not directly from the Lab2\ folder to hopefully make it a bit cleaner. the rest of the files are just from the compilation anyways